+++ agential/cog/agent/reflexion.py
@@ -8,6 +8,7 @@ Paper Repositories:
 
 from typing import Any, Dict, List, Optional, Tuple
 
+import re
 import tiktoken
 
 from langchain.agents.react.base import DocstoreExplorer
@@ -42,7 +43,30 @@ from agential.cog.prompts.agents.reflexion import (
     REFLEXION_REACT_REFLECT_INSTRUCTION,
 )
 from agential.cog.prompts.benchmarks.hotpotqa import HOTPOTQA_FEWSHOT_EXAMPLES_REACT
-from agential.utils.parse import parse_action, remove_newline
+from agential.utils.parse import remove_newline
+
+
+def parse_action(string: str) -> Tuple[str, str]:
+    """Parses an action string into an action type and its argument.
+
+    This method is used in ReAct and Reflexion.
+
+    Args:
+        string (str): The action string to be parsed.
+
+    Returns:
+        Tuple[str, str]: A tuple containing the action type and argument.
+    """
+    pattern = r"^(\w+)\[(.+)\]$"
+    match = re.match(pattern, string)
+
+    if match:
+        action_type = match.group(1)
+        argument = match.group(2)
+    else:  # TODO: Handle parsing/data validation.
+        action_type = ""
+        argument = ""
+    return action_type, argument
 
 
 class ReflexionCoTAgent(BaseAgent):
