+++ agential/cog/strategies/react/qa.py
@@ -2,6 +2,7 @@
 
 from typing import Any, Dict, Tuple
 
+import re
 import tiktoken
 
 from langchain.agents.react.base import DocstoreExplorer
@@ -11,7 +12,30 @@ from tiktoken.core import Encoding
 
 from agential.cog.functional.react import _is_halted, _prompt_agent
 from agential.cog.strategies.react.base import ReActBaseStrategy
-from agential.utils.parse import parse_action, remove_newline
+from agential.utils.parse import remove_newline
+
+
+def parse_qa_action(string: str) -> Tuple[str, str]:
+    """Parses an action string into an action type and its argument.
+
+    This method is used in ReAct and Reflexion.
+
+    Args:
+        string (str): The action string to be parsed.
+
+    Returns:
+        Tuple[str, str]: A tuple containing the action type and argument.
+    """
+    pattern = r"^(\w+)\[(.+)\]$"
+    match = re.match(pattern, string)
+
+    if match:
+        action_type = match.group(1)
+        argument = match.group(2)
+    else:
+        action_type = ""
+        argument = ""
+    return action_type, argument
 
 
 class ReActQAStrategy(ReActBaseStrategy):
@@ -113,7 +137,7 @@ class ReActQAStrategy(ReActBaseStrategy):
         )
         action = remove_newline(action).split("Observation")[0]
         self._scratchpad += " " + action
-        action_type, query = parse_action(action)
+        action_type, query = parse_qa_action(action)
 
         return action_type, query
