+++ agential/utils/parse.py
@@ -79,29 +79,6 @@ def remove_name(text: str, name: str) -> str:
     return lines
 
 
-def parse_action(string: str) -> Tuple[str, str]:
-    """Parses an action string into an action type and its argument.
-
-    This method is used in ReAct and Reflexion.
-
-    Args:
-        string (str): The action string to be parsed.
-
-    Returns:
-        Tuple[str, str]: A tuple containing the action type and argument.
-    """
-    pattern = r"^(\w+)\[(.+)\]$"
-    match = re.match(pattern, string)
-
-    if match:
-        action_type = match.group(1)
-        argument = match.group(2)
-    else:  # TODO: Handle parsing/data validation.
-        action_type = ""
-        argument = ""
-    return action_type, argument
-
-
 def remove_newline(step: str) -> str:
     """Formats a step string by stripping leading/trailing newlines and spaces, and replacing internal newlines with empty space.
